refactor: change fragment generator strategies folder structure and file names
