+++ fragments/fragments-generator/src/strategies/create-box/index.ts
@@ -1,0 +1,1 @@
+export * from './create-box.strategy.js'
