+++ fragments/fragments-generator/src/strategies/create-button/index.ts
@@ -1,0 +1,1 @@
+export * from './create-button.strategy.js'
