+++ fragments/fragments-generator/src/strategies/create-input/index.ts
@@ -1,0 +1,1 @@
+export * from './create-input.strategy.js'
