+++ fragments/fragments-generator/src/strategies/create-text/index.ts
@@ -1,0 +1,1 @@
+export * from './create-text.strategy.js'
