+++ fragments/fragments-generator/src/strategies/strategies.interfaces.ts
@@ -1,4 +1,4 @@
-import { ReactElement } from 'react'
+import type { ReactElement } from 'react'
 
 export interface TreeElement {
   childrenIds: Array<string>
