+++ fragments/fragments-generator/src/strategies/theme-mapping/index.ts
@@ -1,0 +1,1 @@
+export * from './theme-mapping.strategy.js'
