--- fragments/fragments-generator/src/strategy/strategy.constants.ts
@@ -1,3 +1,0 @@
-export const THEME_KEY_PREFIX = '$'
-
-export const colorsIgnorePatterns = ['button', 'input']
